+++ CloudySkies/ModEntry.Commands.cs
@@ -2,13 +2,14 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
+using Leclair.Stardew.CloudySkies.Patches;
 using Leclair.Stardew.Common;
 using Leclair.Stardew.Common.Events;
 
 using StardewModdingAPI;
 
 using StardewValley;
-using StardewValley.BellsAndWhistles;
+using StardewValley.Extensions;
 using StardewValley.Network;
 
 namespace Leclair.Stardew.CloudySkies;
@@ -20,6 +21,7 @@ public partial class ModEntry {
 		UncacheLayers(null, true);
 		Log($"Invalidated weather cache.", LogLevel.Info);
 	}
+
 
 	[ConsoleCommand("cs_history", "View the recorded weather history.")]
 	public void HistoryCommand(string name, string[] args) {
@@ -143,6 +145,7 @@ public partial class ModEntry {
 		Log(sb.ToString(), LogLevel.Info);
 	}
 
+
 	[ConsoleCommand("cs_tomorrow", "Force tomorrow's weather to have a specific type in your current location.")]
 	public void TomorrowCommand(string name, string[] args) {
 		if (!Context.IsWorldReady) {
@@ -166,6 +169,7 @@ public partial class ModEntry {
 
 		UseWeatherTotem(Game1.player, input, item, bypassChecks: true);
 	}
+
 
 	[ConsoleCommand("cs_fix_green_rain", "Remove lingering green rain effects from maps.")]
 	public void FixGreenRainCommand(string name, string[] args) {
@@ -202,6 +206,101 @@ public partial class ModEntry {
 	}
 
 
+	internal static readonly HashSet<string> VANILLA_WEATHER = [
+		"Sun",
+		"Rain",
+		"Storm",
+		"GreenRain",
+		"Snow",
+		"Wind"
+	];
+
+	[ConsoleCommand("cs_set", "Set the current weather in a location context.")]
+	public void SetCommand(string name, string[] args) {
+
+		if (!Context.IsWorldReady) {
+			Log($"Load the game first.", LogLevel.Error);
+			return;
+		}
+
+		if (!Game1.IsMasterGame) {
+			Log($"Only the host can do this.", LogLevel.Error);
+			return;
+		}
+
+		HashSet<TargetLocationContext> targets = [];
+		string? weatherId = null;
+
+		var parser = ArgumentParser.New()
+			.Add<IEnumerable<TargetLocationContext>>("-t", "--target", val => targets.AddRange(val))
+			.AllowMultiple()
+			.AddPositional<string>("WeatherId", val => weatherId = val)
+			.IsRequired()
+			.IsFinal();
+
+		if (!parser.TryParse(args, out string? error)) {
+			Log(error, LogLevel.Error);
+			return;
+		}
+
+		weatherId ??= "Sun";
+
+		if (!VANILLA_WEATHER.Contains(weatherId) && !TryGetWeather(weatherId, out var weatherData)) {
+			Log($"Invalid weather id '{weatherId}'.", LogLevel.Warn);
+			return;
+		}
+
+		if (targets.Count == 0)
+			targets.Add(new TargetLocationContext(Game1.currentLocation?.GetLocationContextId() ?? "Default"));
+
+		Game1_Patches.SaveGreenRainHistory();
+		int total = 0;
+		int changed = 0;
+
+		bool wasDebris = IsDebris;
+
+		foreach (var target in targets) {
+			if (!target.IsValid) {
+				Log($"Skipping invalid location context '{target.Key}'.", LogLevel.Warn);
+				continue;
+			}
+
+			total++;
+
+			var weather = Game1.netWorldState.Value.GetWeatherForLocation(target.Key);
+			if (weather.Weather == weatherId)
+				continue;
+
+			changed++;
+
+			weather.Weather = weatherId;
+			weather.UpdateFromWeatherIncludingVanilla(true);
+
+			// Is this the default context?
+			if (target.Key == "Default") {
+				Game1.isRaining = weather.IsRaining;
+				Game1.isSnowing = weather.IsSnowing;
+				Game1.isLightning = weather.IsLightning;
+				Game1.isDebrisWeather = weather.IsDebrisWeather;
+				Game1.isGreenRain = weather.IsGreenRain;
+			}
+		}
+
+		Game1_Patches.UpdateGreenRainLocations();
+
+		if (IsDebris && !wasDebris)
+			Game1.populateDebrisWeatherArray();
+
+		Game1.updateWeatherIcon();
+		if (Game1.currentLocation != null)
+			GameLocation.HandleMusicChange(Game1.currentLocation, Game1.currentLocation);
+
+		Log($"Changed weather in {changed} of {total} matching location contexts.", LogLevel.Info);
+	}
+
+	internal static bool IsDebris => Game1.isDebrisWeather && Game1.currentLocation != null && Game1.currentLocation.IsOutdoors && !Game1.currentLocation.ignoreDebrisWeather.Value;
+
+
 	[ConsoleCommand("cs_reset_critters", "Clear all the existing critters on the current map, and then spawn new ones.")]
 	public void ResetCrittersCommand(string name, string[] args) {
 		if (!Context.IsWorldReady) {
@@ -210,7 +309,7 @@ public partial class ModEntry {
 		}
 
 		var loc = Game1.currentLocation;
-		if (loc is null || !loc.IsOutdoors) { 
+		if (loc is null || !loc.IsOutdoors) {
 			Log($"Can only be used outdoors.", LogLevel.Warn);
 			return;
 		}