+++ CloudySkies/Triggers/SpawnArtifactSpots.cs
@@ -200,8 +200,8 @@ public static partial class Triggers {
 		if (toSpawn is null)
 			return false;
 
-		Vector2 tilePos = new Vector2(x * 64, y * 64);
-		ItemQueryContext ctx = new(location, who, Game1.random);
+		Vector2 tilePos = new(x * 64, y * 64);
+		ItemQueryContext ctx = new(location, who, Game1.random, sourcePhrase: "custom artifact spot");
 
 		Item item = ItemQueryResolver.TryResolveRandomItem(toSpawn, ctx, logError: (query, error) => {
 			Instance.Log($"Error parsing item query '{query}' for artifact spot: {error}", LogLevel.Error);